panel: address Ben's feedback
