+++ src/vs/workbench/parts/debug/browser/repl.ts
@@ -17,7 +17,8 @@ import viewer = require('vs/workbench/parts/debug/browser/replViewer');
 import debug = require('vs/workbench/parts/debug/common/debug');
 import debugactions = require('vs/workbench/parts/debug/electron-browser/debugActions');
 import replhistory = require('vs/workbench/parts/debug/common/replHistory');
-import { Panel, ClosePanelAction } from 'vs/workbench/browser/panel';
+import { Panel } from 'vs/workbench/browser/panel';
+import { ClosePanelAction } from 'vs/workbench/browser/parts/panel/panelPart';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IContextViewService, IContextMenuService } from 'vs/platform/contextview/browser/contextView';
 import { IInstantiationService, INullService } from 'vs/platform/instantiation/common/instantiation';
